+++ helpers/build.js
@@ -4,10 +4,10 @@ const sass = require('./builders/sass')
 const scss = require('./builders/scss')
 const image = require('./builders/image')
 
-module.exports = async () => {
+module.exports = (async () => {
   await js()
   await sass()
   await scss()
   await image()
   await pug()
-}
+})()