+++ web/js/index.js
@@ -1,0 +1,10 @@
+import Vue from 'vue'
+import router from '../router/index.js'
+
+const APP = new Vue({
+  router,
+  name: 'APP',
+  el: '#app'
+})
+
+export default APP